Change the CapabilitiesFlags property into booleans
Each bit in CapabilitiesFlags dbus property is a chassis
capability. Setting a capability overwrites other capabilities.
Read-and-update of the property results in race condition/confusion.
To tackle this, added a boolean property for each of chassis
capability. The new properties will be read and updated in ipmi
command handling code.

The capabilities listed below are supported by Intel and so enabled
the corresponding bits/properties by default.
1. Intrusion sensor (bit 0)
2. Front panel lockout (bit 1)
3. NMI (bit 2)

The original CapabilitiesFlags property will be deprecated once the
changes are made to use new booleans in all the repos.

Tested:
------

root@intel-obmc:~# ipmitool raw 0x00 0x00
 07 22 44 66 88 aa

Change-Id: I7f9ec00c1bf94d550dd000634011ef8e7a460967
Signed-off-by: Karthick Sundarrajan <karthick.sundarrajan@intel.com>