Add a configuration option to allow reset on MCERR (#14)
MCERR is another type of error that can be recovered by a reset. This
adds a configuration option to control if the BMC will allow that
recovery.

Tested:
Injected an MCERR and confirmed that when this option is true, recovery
is enabled. When this option is false, recovery is blocked.

Signed-off-by: Jason M. Bills <jason.m.bills@intel.com>